chore: move renderers to shared folder
