+++ unime/src/routes/credentials/[id]/(renderers)/DataUrlImageRenderer.svelte
